+++ frontend/src/components/Comments/Comments.scss
@@ -1,7 +1,7 @@
 .comments {
   position: relative;
   margin: 0 auto;
-  width: 144rem;
+  width: 100%;
   background: #e5f0ff;
   padding: 7.2rem 10rem;
   box-sizing: border-box;
@@ -30,8 +30,14 @@
     letter-spacing: 0em;
   }
 
+  &__go_position {
+    text-align: center;
+    display: flex;
+    justify-content: center;
+    margin-top: 3.2rem;
+  }
+
   &__list {
-    list-style: none;
     padding: 0 0 7.2rem;
     margin: 0;
     display: flex;
