+++ frontend/src/components/CommentsAbout/CommentsAbout.scss
@@ -1,0 +1,19 @@
+.comments {
+  &__about {
+    display: grid;
+    padding: 0 0 7.2rem;
+    margin: 0;
+    grid-template-columns: repeat(4, 1fr);
+    grid-template-rows: repeat(2, 1fr);
+    gap: 3.2rem;
+  }
+  &__show {
+    font-weight: 700;
+    font-size: 16px;
+    color: rgba(22, 59, 108, 1);
+    background: none;
+    border: none;
+    margin: 0 auto;
+    display: block;
+  }
+}
