+++ frontend/src/utils/commentsData.js
@@ -1,0 +1,110 @@
+export const comments = [
+  {
+    id: 1,
+    rating: 5.0,
+    comment:
+      'С самого начала, когда я обратилась к ним, я ощутила заботу о клиентах. Команда всегда приходит вовремя. Что меня особо поражает, это их тщательность и внимание к деталям. Они не только делают качественную уборку, но и обращают внимание на мелочи, которые я раньше не замечала. Каждый раз я приятно удивляюсь насколько свежо и чисто в моей квартире. Кажется, что они каждый раз врубают режим магии, потому что результат превосходит все мои ожидания.',
+    clientName: 'Ольга',
+    charwomanName: 'Светлана',
+  },
+  {
+    id: 2,
+    rating: 5,
+    comment: 'Никогда не видел такой тщательной уборки. Очень доволен результатом',
+    clientName: 'Дмитрий',
+    charwomanName: 'Светлана',
+  },
+  {
+    id: 3,
+    rating: 5,
+    comment: 'Сервис по уборке квартир превзошел все мои ожидания. Супер дружелюбные и эффективные работники!',
+    clientName: 'Наталья',
+    charwomanName: 'Оксана',
+  },
+  {
+    id: 4,
+    rating: 5,
+    comment: 'Большое спасибо за обновленный и свежий вид моей квартиры. Постоянно буду обращаться за помощью',
+    clientName: 'Юлия',
+    charwomanName: 'Ирина',
+  },
+  {
+    id: 5,
+    rating: 5,
+    comment: 'Работники сервиса по уборке квартир очень внимательны к деталям. Все было прекрасно убрано и аккуратно.',
+    clientName: 'Елена',
+    charwomanName: 'Оксана',
+  },
+  {
+    id: 6,
+    rating: 5.0,
+    comment:
+      'С самого начала, когда я обратилась к ним, я ощутила заботу о клиентах. Команда всегда приходит вовремя. Что меня особо поражает, это их тщательность и внимание к деталям. Они не только делают качественную уборку, но и обращают внимание на мелочи, которые я раньше не замечала. Каждый раз я приятно удивляюсь насколько свежо и чисто в моей квартире. Кажется, что они каждый раз врубают режим магии, потому что результат превосходит все мои ожидания.',
+    clientName: 'Ольга',
+    charwomanName: 'Светлана',
+  },
+  {
+    id: 7,
+    rating: 5,
+    comment: 'Никогда не видел такой тщательной уборки. Очень доволен результатом',
+    clientName: 'Дмитрий',
+    charwomanName: 'Светлана',
+  },
+  {
+    id: 8,
+    rating: 5,
+    comment: 'Сервис по уборке квартир превзошел все мои ожидания. Супер дружелюбные и эффективные работники!',
+    clientName: 'Наталья',
+    charwomanName: 'Оксана',
+  },
+  {
+    id: 9,
+    rating: 5,
+    comment: 'Большое спасибо за обновленный и свежий вид моей квартиры. Постоянно буду обращаться за помощью',
+    clientName: 'Юлия',
+    charwomanName: 'Ирина',
+  },
+  {
+    id: 10,
+    rating: 5,
+    comment: 'Работники сервиса по уборке квартир очень внимательны к деталям. Все было прекрасно убрано и аккуратно.',
+    clientName: 'Елена',
+    charwomanName: 'Оксана',
+  },
+  {
+    id: 11,
+    rating: 5.0,
+    comment:
+      'С самого начала, когда я обратилась к ним, я ощутила заботу о клиентах. Команда всегда приходит вовремя. Что меня особо поражает, это их тщательность и внимание к деталям. Они не только делают качественную уборку, но и обращают внимание на мелочи, которые я раньше не замечала. Каждый раз я приятно удивляюсь насколько свежо и чисто в моей квартире. Кажется, что они каждый раз врубают режим магии, потому что результат превосходит все мои ожидания.',
+    clientName: 'Ольга',
+    charwomanName: 'Светлана',
+  },
+  {
+    id: 12,
+    rating: 5,
+    comment: 'Никогда не видел такой тщательной уборки. Очень доволен результатом',
+    clientName: 'Дмитрий',
+    charwomanName: 'Светлана',
+  },
+  {
+    id: 13,
+    rating: 5,
+    comment: 'Сервис по уборке квартир превзошел все мои ожидания. Супер дружелюбные и эффективные работники!',
+    clientName: 'Наталья',
+    charwomanName: 'Оксана',
+  },
+  {
+    id: 14,
+    rating: 5,
+    comment: 'Большое спасибо за обновленный и свежий вид моей квартиры. Постоянно буду обращаться за помощью',
+    clientName: 'Юлия',
+    charwomanName: 'Ирина',
+  },
+  {
+    id: 15,
+    rating: 5,
+    comment: 'Работники сервиса по уборке квартир очень внимательны к деталям. Все было прекрасно убрано и аккуратно.',
+    clientName: 'Елена',
+    charwomanName: 'Оксана',
+  },
+]
